--- reproductions/README.md
@@ -1,9 +1,0 @@
-# uniforms-reproductions
-
-Codesandbox reproduction acceleration.
-
-## running
-
-```
-npm start
-```
