--- reproductions/schema/index.js
@@ -1,3 +1,0 @@
-export { default as jsonSchema } from './json-schema';
-export { default as graphqlSchema } from './graphql-schema';
-export { default as simpleSchema2 } from './simple-schema-2';
