+++ reproductions/schema/json-schema.js
@@ -1,3 +1,18 @@
+import Ajv from 'ajv';
+import { JSONSchemaBridge } from 'uniforms-bridge-json-schema';
+
+const ajv = new Ajv({ allErrors: true, useDefaults: true });
+
+function createValidator(schema) {
+  const validator = ajv.compile(schema);
+  return model => {
+    validator(model);
+    if (validator.errors && validator.errors.length) {
+      throw { details: validator.errors };
+    }
+  };
+}
+
 const schema = {
   type: 'object',
   properties: {
@@ -37,4 +52,6 @@ const schema = {
   required: ['b', 'c', 'd', 'e']
 };
 
-export default schema;
+const schemaValidator = createValidator(schema);
+
+export default new JSONSchemaBridge(schema, schemaValidator);
