--- reproductions/schema/json-schema/index.js
@@ -1,8 +1,0 @@
-import { JSONSchemaBridge } from 'uniforms-bridge-json-schema';
-
-import { createValidator } from './validator';
-import schema from './schema';
-
-const schemaValidator = createValidator(schema);
-
-export default new JSONSchemaBridge(schema, schemaValidator);
